+++ Content.Shared/Singularity/EntitySystems/SharedEmitterSystem.cs
@@ -1,0 +1,6 @@
+﻿namespace Content.Shared.Singularity.EntitySystems;
+
+public abstract class SharedEmitterSystem : EntitySystem
+{
+
+}
